Remove duplicate people from check Camdram show page
It is possible that the same person might be involved in multiple roles
on the same show, in which case they will be duplicated on the 'check
Camdram show' page. This commit removes these duplicate tuples by
filtering for unique tuples based on the Camdram ID of the person
involved.

Signed-off-by: Charlie Jonas <charlie@charliejonas.co.uk>